+++ src/resources/index.ts
@@ -1,25 +1,24 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
+export { Audio, type AudioGenerateResponse, type AudioGenerateParams } from './audio';
 export { Auth } from './auth/auth';
+export { Chat, type ChatCreateCompletionResponse, type ChatCreateCompletionParams } from './chat';
 export { Data, type DataIngestResponse, type DataIngestParams } from './data/data';
 export { Health, type HealthCheckResponse } from './health';
+export { Images, type ImageGenerateResponse, type ImageGenerateParams } from './images';
 export {
   Inference,
   type InferenceGenerateCompletionResponse,
   type InferenceGenerateMultimodalityCompletionResponse,
-  type InferenceGeneratePersonaChatResponse,
   type InferenceGenerateCompletionParams,
   type InferenceGenerateMultimodalityCompletionParams,
-  type InferenceGeneratePersonaChatParams,
 } from './inference';
 export {
   Modal,
   type ModalLearnResponse,
   type ModalQueryResponse,
-  type ModalQueryMultimodalityResponse,
   type ModalLearnParams,
   type ModalQueryParams,
-  type ModalQueryMultimodalityParams,
 } from './modal';
 export {
   Personas,
@@ -36,4 +35,6 @@ export {
   type ProjectDeleteResponse,
   type ProjectCreateParams,
 } from './projects';
+export { Text, type TextGenerateResponse, type TextGenerateParams } from './text';
 export { Users, type UserCreateOrGetResponse, type UserCreateOrGetParams } from './users';
+export { Video, type VideoGenerateResponse, type VideoGenerateParams } from './video';